*: clean up the usage of slog (#288)
This commit brings 2 big changes to the API:

1. No more default logger. Users are required to give a logger to use
the most of structs in this library. This makes the interface simple.
2. No more tag scope value. Some APIs insert tag into KVs, some don't.
It's confusing and easily lead to duplicate KV definitions. Users are
required to insert tag into KVs before passing the logger to the
library, and there is no more explicit configuration of tag.